+++ package.json
@@ -13,6 +13,7 @@
   },
   "dependencies": {
     "geojson-coords": "0.0.0",
+    "geojson-flatten": "^0.1.0",
     "minimist": "^1.1.1",
     "rbush": "^1.4.2",
     "tile-reduce": "^3.0.0",
